+++ Packages/com.unity.cloud.gltfast/Runtime/Scripts/GameObjectInstantiator.cs
@@ -186,6 +186,20 @@ namespace GLTFast
         }
 
         /// <inheritdoc />
+        public virtual void CreateNode(
+            uint nodeIndex,
+            uint? parentIndex,
+            Vector3 position,
+            Quaternion rotation,
+            Vector3 scale,
+            string name
+        )
+        {
+            CreateNode(nodeIndex, parentIndex, position, rotation, scale);
+            SetNodeName(nodeIndex, name);
+        }
+
+        /// <inheritdoc />
         public virtual void SetNodeName(uint nodeIndex, string name)
         {
             m_Nodes[nodeIndex].name = name ?? $"Node-{nodeIndex}";